+++ fermata/src/main/java/me/aap/fermata/media/engine/SubtitleStreamInfo.java
@@ -4,7 +4,7 @@ package me.aap.fermata.media.engine;
  * @author Andrey Pavlenko
  */
 public class SubtitleStreamInfo extends MediaStreamInfo {
-	public SubtitleStreamInfo(int id, String language, String description) {
+	public SubtitleStreamInfo(long id, String language, String description) {
 		super(id, language, description);
 	}
 }
